// PayU_Frontend/src/types/process.ts
import type { InvoiceData } from "./invoice";
import type { POData } from "./purchase_order";


export interface Vendor {
  name: string;
  email: string;
  address: string;
  country_code: string;
  mobile_number: string;
  gst_number: string;
  bank_name: string;
  account_holder_name: string;
  account_number: string;
  ifsc_code: string;
}

export type ExtractedFile =
  | {
      id: string;
      fileName: string;
      file: File;
      status: "extracting" | "done" | "error";
      type: "invoice";
      extractedData?: InvoiceData;
    }
  | {
      id: string;
      fileName: string;
      file: File;
      status: "extracting" | "done" | "error";
      type: "po";
      extractedData?: POData;
    };